rename driver psu_hm7044
