+++ panduza-py-class-power-supply/__init__.py
@@ -1,10 +1,10 @@
 import imp
-from .driver_hm7044 import DriverHm7044
+from .driver_psu_hm7044 import DriverPsuHm7044
 from .driver_ka005p import DriverKA005P
 from .driver_ql335p import DriverQL335P
 
 PZA_DRIVERS_LIST=[
-    DriverHm7044,
+    DriverPsuHm7044,
     DriverKA005P,
     DriverQL335P
 ]
